+++ packages/config/src/assets.ts
@@ -155,6 +155,11 @@ export const sdn = new Asset({
   originSymbol: 'SDN',
 });
 
+export const soon = new Asset({
+  key: 'soon',
+  originSymbol: 'SOON',
+});
+
 export const teer = new Asset({
   key: 'teer',
   originSymbol: 'TEER',
@@ -214,6 +219,7 @@ export const assetsList: Asset[] = [
   sdn,
   teer,
   tt1,
+  soon,
   unit,
   usdt,
   xrt,