+++ packages/config/src/configs/index.ts
@@ -35,6 +35,7 @@ import { shidenConfig } from './shiden';
 import { statemineConfig } from './statemine';
 import { statemineAlphanetConfig } from './statemineAlphanet';
 import { statemintConfig } from './statemint';
+import { subsocialAlphanetConfig } from './subsocialAlphanet';
 import { uniqueAlphaConfig } from './uniqueAlpha';
 
 export const chainsConfigList: ChainConfig[] = [
@@ -73,6 +74,7 @@ export const chainsConfigList: ChainConfig[] = [
   statemineAlphanetConfig,
   statemineConfig,
   statemintConfig,
+  subsocialAlphanetConfig,
   uniqueAlphaConfig,
 ];