+++ packages/config/src/configs/moonbaseAlpha.ts
@@ -1,5 +1,5 @@
 import { BalanceBuilder, ContractBuilder } from '@moonbeam-network/xcm-builder';
-import { dev, eq, eqd, lit, nodl, paring, tt1, unit } from '../assets';
+import { dev, eq, eqd, lit, nodl, paring, soon, tt1, unit } from '../assets';
 import {
   alphanetRelay,
   darwiniaPangoro,
@@ -8,6 +8,7 @@ import {
   moonbaseAlpha,
   nodleEden,
   statemineAlphanet,
+  subsocialAlphanet,
 } from '../chains';
 import { AssetConfig } from '../types/AssetConfig';
 import { ChainConfig } from '../types/ChainConfig';
@@ -103,6 +104,20 @@ export const moonbaseAlphaConfig = new ChainConfig({
       destinationFee: {
         amount: 4,
         asset: paring,
+      },
+      fee: {
+        asset: dev,
+        balance: BalanceBuilder().system().account(),
+      },
+    }),
+    new AssetConfig({
+      asset: soon,
+      balance: BalanceBuilder().assets().account(),
+      contract: ContractBuilder().Xtokens().transfer(),
+      destination: subsocialAlphanet,
+      destinationFee: {
+        amount: 0.4,
+        asset: soon,
       },
       fee: {
         asset: dev,